requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.2+cpu
torchvision==0.16.2+cpu
Pillow==10.1.0
numpy==1.24.3
pydantic==2.5.2
python-multipart==0.0.6
